refactor(rpc): extract primary-key operation handlers
Move primary-key read, write, delete, and batch operation handling into primary_key_operations.cc.
Register primary_key_operations.cc in the server build.
