+++ server/CMakeLists.txt
@@ -59,6 +59,7 @@ set(SOURCES
     rpc/flat_plan_encode.hh
     rpc/parallel_scan.cc
     rpc/parallel_scan.hh
+    rpc/primary_key_operations.cc
     rpc/read_plan_executor.cc
     rpc/row_codec.cc
     rpc/row_codec.hh
