+++ server/rpc/lineairdb_rpc.hh
@@ -71,12 +71,16 @@ private:
     void handleTxRead(const std::string& message, std::string& result);
     void handleTxBatchRead(const std::string& message, std::string& result);
     void handleTxBatchWrite(const std::string& message, std::string& result);
-    void handleTxStatelessRead(const std::string& message, std::string& result);
-    void handleTxStatelessBatchRead(const std::string& message, std::string& result);
-    void handleTxExecuteReadPlan(const std::string& message, std::string& result);
-    void handleTxValidateAndCommit(const std::string& message, std::string& result);
     void handleTxWrite(const std::string& message, std::string& result);
     void handleTxDelete(const std::string& message, std::string& result);
+
+    // Stateless operations
+    void handleTxStatelessRead(const std::string& message, std::string& result);
+    void handleTxStatelessBatchRead(const std::string& message, std::string& result);
+    void handleTxValidateAndCommit(const std::string& message, std::string& result);
+
+    // Read-plan execution
+    void handleTxExecuteReadPlan(const std::string& message, std::string& result);
 
     // Secondary index operations
     void handleTxReadSecondaryIndex(const std::string& message, std::string& result);
