+++ server/rpc/primary_key_operations.cc
@@ -1,0 +1,214 @@
+#include "lineairdb_rpc.hh"
+
+#include <cstddef>
+#include <string>
+
+#include "../../common/log.h"
+#include "lineairdb.pb.h"
+
+// Primary-key row read/write RPC handlers. TxBatchWrite also carries the
+// proxy's mixed write envelope, including secondary-index batch variants.
+
+void LineairDBRpc::handleTxRead(const std::string& message,
+                                std::string& result) {
+    LOG_DEBUG("Handling TxRead");
+
+    LineairDB::Protocol::TxRead::Request request;
+    LineairDB::Protocol::TxRead::Response response;
+
+    request.ParseFromString(message);
+
+    const int64_t tx_id = request.transaction_id();
+    auto* tx = tx_manager_->get_transaction(tx_id);
+    if (tx) {
+        if (!request.table_name().empty()) {
+            tx->SetTable(request.table_name());
+        }
+        auto read_result = tx->Read(request.key());
+        response.set_is_aborted(tx->IsAborted());
+
+        if (read_result.first != nullptr) {
+            response.set_found(true);
+            std::string value(reinterpret_cast<const char*>(read_result.first),
+                              read_result.second);
+            response.set_value(value);
+        } else {
+            response.set_found(false);
+        }
+
+        LOG_DEBUG("Read key '%s' from transaction %ld: %s",
+                  request.key().c_str(), tx_id,
+                  read_result.first != nullptr ? "found" : "not found");
+    } else {
+        response.set_found(false);
+        response.set_is_aborted(true);
+        LOG_WARNING("Transaction not found for read: %ld", tx_id);
+    }
+
+    result = response.SerializeAsString();
+}
+
+void LineairDBRpc::handleTxBatchRead(const std::string& message,
+                                     std::string& result) {
+    LineairDB::Protocol::TxBatchRead::Request request;
+    LineairDB::Protocol::TxBatchRead::Response response;
+
+    request.ParseFromString(message);
+
+    const int64_t tx_id = request.transaction_id();
+    auto* tx = tx_manager_->get_transaction(tx_id);
+    if (tx) {
+        if (!request.table_name().empty()) {
+            tx->SetTable(request.table_name());
+        }
+        for (int i = 0; i < request.keys_size(); i++) {
+            auto* read_result = response.add_results();
+            auto pair = tx->Read(request.keys(i));
+            if (pair.first != nullptr) {
+                read_result->set_found(true);
+                read_result->set_value(
+                    reinterpret_cast<const char*>(pair.first), pair.second);
+            } else {
+                read_result->set_found(false);
+            }
+        }
+        response.set_is_aborted(tx->IsAborted());
+    } else {
+        response.set_is_aborted(true);
+        LOG_WARNING("Transaction not found for batch_read: %ld", tx_id);
+    }
+
+    result = response.SerializeAsString();
+}
+
+void LineairDBRpc::handleTxBatchWrite(const std::string& message,
+                                      std::string& result) {
+    LineairDB::Protocol::TxBatchWrite::Request request;
+    LineairDB::Protocol::TxBatchWrite::Response response;
+
+    request.ParseFromString(message);
+
+    const int64_t tx_id = request.transaction_id();
+    auto* tx = tx_manager_->get_transaction(tx_id);
+    if (tx) {
+        if (!request.table_name().empty()) {
+            tx->SetTable(request.table_name());
+        }
+
+        if (!tx->IsAborted()) {
+            for (int i = 0; i < request.ops_size(); i++) {
+                const auto& op = request.ops(i);
+                const std::string& op_table =
+                    op.table_name().empty() ? request.table_name()
+                                            : op.table_name();
+                if (!op_table.empty()) {
+                    tx->SetTable(op_table);
+                }
+                switch (op.type()) {
+                    case LineairDB::Protocol::BATCH_OP_WRITE: {
+                        const std::string& value_str = op.value();
+                        tx->Write(
+                            op.key(),
+                            reinterpret_cast<const std::byte*>(
+                                value_str.c_str()),
+                            value_str.size());
+                        break;
+                    }
+                    case LineairDB::Protocol::BATCH_OP_DELETE:
+                        tx->Delete(op.key());
+                        break;
+                    case LineairDB::Protocol::BATCH_OP_SECONDARY_INDEX_WRITE: {
+                        const std::string& pk = op.primary_key();
+                        tx->WriteSecondaryIndex(
+                            op.index_name(), op.secondary_key(),
+                            reinterpret_cast<const std::byte*>(pk.c_str()),
+                            pk.size());
+                        break;
+                    }
+                    case LineairDB::Protocol::BATCH_OP_SECONDARY_INDEX_DELETE: {
+                        const std::string& pk = op.primary_key();
+                        tx->DeleteSecondaryIndex(
+                            op.index_name(), op.secondary_key(),
+                            reinterpret_cast<const std::byte*>(pk.c_str()),
+                            pk.size());
+                        break;
+                    }
+                    case LineairDB::Protocol::BATCH_OP_UNKNOWN:
+                    default:
+                        break;
+                }
+                if (tx->IsAborted()) break;
+            }
+        }
+
+        response.set_success(!tx->IsAborted());
+        response.set_is_aborted(tx->IsAborted());
+    } else {
+        response.set_success(false);
+        response.set_is_aborted(true);
+        LOG_WARNING("Transaction not found for batch_write: %ld", tx_id);
+    }
+
+    result = response.SerializeAsString();
+}
+
+void LineairDBRpc::handleTxWrite(const std::string& message,
+                                 std::string& result) {
+    LOG_DEBUG("Handling TxWrite");
+
+    LineairDB::Protocol::TxWrite::Request request;
+    LineairDB::Protocol::TxWrite::Response response;
+
+    request.ParseFromString(message);
+
+    const int64_t tx_id = request.transaction_id();
+    auto* tx = tx_manager_->get_transaction(tx_id);
+    if (tx) {
+        if (!request.table_name().empty()) {
+            tx->SetTable(request.table_name());
+        }
+        const std::string& value_str = request.value();
+        tx->Write(request.key(),
+                  reinterpret_cast<const std::byte*>(value_str.c_str()),
+                  value_str.size());
+        response.set_is_aborted(tx->IsAborted());
+        response.set_success(!tx->IsAborted());
+        LOG_DEBUG("Wrote key '%s' to transaction %ld", request.key().c_str(),
+                  tx_id);
+    } else {
+        response.set_success(false);
+        response.set_is_aborted(true);
+        LOG_WARNING("Transaction not found for write: %ld", tx_id);
+    }
+
+    result = response.SerializeAsString();
+}
+
+void LineairDBRpc::handleTxDelete(const std::string& message,
+                                  std::string& result) {
+    LOG_DEBUG("Handling TxDelete");
+
+    LineairDB::Protocol::TxDelete::Request request;
+    LineairDB::Protocol::TxDelete::Response response;
+
+    request.ParseFromString(message);
+
+    const int64_t tx_id = request.transaction_id();
+    auto* tx = tx_manager_->get_transaction(tx_id);
+    if (tx) {
+        if (!request.table_name().empty()) {
+            tx->SetTable(request.table_name());
+        }
+        tx->Delete(request.key());
+        response.set_is_aborted(tx->IsAborted());
+        response.set_success(!tx->IsAborted());
+        LOG_DEBUG("Deleted key '%s' from transaction %ld",
+                  request.key().c_str(), tx_id);
+    } else {
+        response.set_success(false);
+        response.set_is_aborted(true);
+        LOG_WARNING("Transaction not found for delete: %ld", tx_id);
+    }
+
+    result = response.SerializeAsString();
+}
